timeclock/test_timeclock.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Vibrator</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="doTest()">

<script class="testbody" type="text/javascript;version=1.7">

var nav = SpecialPowers.wrap(navigator);
var currentDate;

function doTest() {
try {
  navigator.mozTime;
  ok(false, "navigator.mozTime should throw in regular mochitest");
} catch(e) {
  ok(true, "navigator.mozTime should throw in regular mochitest");
}

try {
  nav.mozTime.set(null);
  ok(false, "chromeWindow.navigator.mozTime.set(null) should throw");
} catch(e) {
  ok(true, "chromeWindow.navigator.mozTime.set(null) should throw");
}
try {
  nav.mozTime.set({});
  ok(false, "chromeWindow.navigator.mozTime.set({}) should throw");
} catch(e) {
  ok(true, "chromeWindow.navigator.mozTime.set({}) should throw");
}

currentDate = Date.now();
window.addEventListener('moztimechange', doNext,true);
nav.mozTime.set(0);
}

var counter = 0;
function doNext() {
if (counter = 1) {
  is(currentDate, Date.now(), 'Date has succesfully changed to' + Date.now());
  SimpleTest.finish();
}
ok(true, 'Date has changed to' + Date.now());
counter++;
nav.mozTime.set(currentDate);
}

SimpleTest.waitForExplicitFinish();

</script>
</body>

</html>
